+++ src/manager/explorer/conanCache.ts
@@ -82,11 +82,18 @@ export class ConanCacheExplorerManager extends ExtensionManager {
      * Callback method if a recipe item in the treeview in selected
      */
     private recipeItemSelected() {
-        this.nodeProviderConanRecipe.setSelectedRecipe(this.treeViewConanRecipe.selection[0].label);
-
-        // Change the title of the treeview for package explorer to match the selected recipe
-        this.treeViewConanPackage.title = this.treeViewConanRecipe.selection[0].label;
-        this.nodeProviderConanPackage.refresh(this.treeViewConanRecipe.selection[0].label, this.context.workspaceState.get("show-dirty")!);
+        // Only works with non editable package
+        if (this.treeViewConanRecipe.selection[0].isEditable()) {
+            this.nodeProviderConanPackage.refresh("", this.context.workspaceState.get("show-dirty")!); // Empty the binary package treeview
+            this.treeViewConanPackage.title = "Conan - Package"; // Reset the title of the binary package treeview panel
+        }
+        else {
+             this.nodeProviderConanRecipe.setSelectedRecipe(this.treeViewConanRecipe.selection[0].label);
+    
+            // Change the title of the treeview for package explorer to match the selected recipe
+            this.treeViewConanPackage.title = this.treeViewConanRecipe.selection[0].label;
+            this.nodeProviderConanPackage.refresh(this.treeViewConanRecipe.selection[0].label, this.context.workspaceState.get("show-dirty")!);
+        }
     }
 
     /**
@@ -120,7 +127,12 @@ export class ConanCacheExplorerManager extends ExtensionManager {
      */
     private recipeOpenExplorer(node: ConanRecipeItem) {
         try {
-            vscode.commands.executeCommand('revealFileInOS', vscode.Uri.file(this.conanApi.getRecipePath(node.label)!));
+            if (node.isEditable()) {
+                vscode.commands.executeCommand('revealFileInOS', vscode.Uri.file(node.model.path));
+            }
+            else {
+                vscode.commands.executeCommand('revealFileInOS', vscode.Uri.file(this.conanApi.getRecipePath(node.label)!));
+            }
         }
         catch (err) {
             vscode.window.showErrorMessage((err as Error).message);
@@ -133,8 +145,15 @@ export class ConanCacheExplorerManager extends ExtensionManager {
      */
     private recipeOpenVSCode(node: ConanRecipeItem) {
         try {
-            let packagePath = this.conanApi.getRecipePath(node.label);
-            vscode.commands.executeCommand('vscode.openFolder', vscode.Uri.file(packagePath!), true);
+            if (node.isEditable()) {
+                // The path in the model is referring to the conanfile.py
+                // We want to open the parent path of the file
+                vscode.commands.executeCommand('vscode.openFolder', vscode.Uri.file(node.model.path + "/.."), true);
+            }
+            else {
+                let packagePath = this.conanApi.getRecipePath(node.label);
+                vscode.commands.executeCommand('vscode.openFolder', vscode.Uri.file(packagePath!), true);
+            }
         }
         catch (err) {
             vscode.window.showErrorMessage((err as Error).message);
@@ -147,7 +166,12 @@ export class ConanCacheExplorerManager extends ExtensionManager {
      */
     private recipeRemove(node: ConanRecipeItem) {
         try {
-            vscode.window
+            if (node.isEditable()) {
+                this.conanApi.removeEditablePackageRecipe(node.label);
+                this.nodeProviderConanRecipe.refresh();
+            }
+            else {
+                vscode.window
                 .showWarningMessage(`Are you sure you want to remove the recipe '${node.label}'?`, ...["Yes", "No"])
                 .then((answer) => {
                     if (answer === "Yes") {
@@ -158,6 +182,8 @@ export class ConanCacheExplorerManager extends ExtensionManager {
                         this.treeViewConanPackage.title = "Conan - Package"; // Reset the title of the binary package treeview panel
                     }
                 });
+            }
+            
         }
         catch (err) {
             vscode.window.showErrorMessage((err as Error).message);
@@ -166,8 +192,13 @@ export class ConanCacheExplorerManager extends ExtensionManager {
 
     private recipeCopyPathToClipboard(node: ConanRecipeItem) {
         try {
-            let recipePath = this.conanApi.getRecipePath(node.label);
-            vscode.env.clipboard.writeText(recipePath!);
+            if (node.isEditable()) {
+                vscode.env.clipboard.writeText(node.model.path);
+            }
+            else {
+                let recipePath = this.conanApi.getRecipePath(node.label);
+                vscode.env.clipboard.writeText(recipePath!);
+            }
         }
         catch {
             vscode.window.showErrorMessage("Unable to copy the path to clipboard");