+++ src/manager/vsconanWorkspace.ts
@@ -49,6 +49,8 @@ export class VSConanWorkspaceManager extends ExtensionManager {
         this.registerCommand("vsconan.conan.source", () => this.executeConanCommand(ConanCommand.source));
         this.registerCommand("vsconan.conan.package", () => this.executeConanCommand(ConanCommand.package));
         this.registerCommand("vsconan.conan.package.export", () => this.executeConanCommand(ConanCommand.packageExport));
+        this.registerCommand("vsconan.conan.editable.add", () => this.addEditablePackage());
+        this.registerCommand("vsconan.conan.editable.remove", () => this.removeEditablePackage());
         this.registerCommand("vsconan.config.global.create", () => this.createGlobalConfig());
         this.registerCommand("vsconan.config.global.open", () => this.openGlobalConfig());
         this.registerCommand("vsconan.config.workspace.create", () => this.createWorkspaceConfig());
@@ -390,5 +392,139 @@ export class VSConanWorkspaceManager extends ExtensionManager {
                 }
             }
         });
+    }
+
+    private async addEditablePackage() {
+        // List of workspace
+        let wsList = vscode.workspace.workspaceFolders;
+
+        let conanWorkspaceList: Array<string> = [];
+
+        if (wsList != undefined) {
+            try {
+                // Filtering workspace into conan workspace
+                // Minimize options for the quickpick
+                for (let ws of wsList!) {
+                    if (utils.conan.isFolderConanProject(ws.uri.fsPath)) {
+                        conanWorkspaceList.push(ws.uri.fsPath);
+                    }
+                }
+
+                // Picking the conan workspace after filtering the workspace into conan workspace
+                const quickPick = vscode.window.createQuickPick<vscode.QuickPickItem>();
+                let quickPickItems = [];
+                for (let i = 0; i < conanWorkspaceList.length; i++) {
+                    quickPickItems.push({
+                        label: conanWorkspaceList[i],
+                        description: "",
+                        detail: "",
+                        index: i
+                    });
+                }
+                quickPickItems.map(label => ({ label }));
+                quickPick.items = quickPickItems;
+
+                const wsChoice = await vscode.window.showQuickPick(quickPickItems);
+
+                if (wsChoice) {
+                    // Get the name and version in the recipe
+                    let name = this.conanApi.getRecipeAttribute(wsChoice!.label, "name");
+                    let version = this.conanApi.getRecipeAttribute(wsChoice!.label, "version");
+                    let packageInformation = `${name}/${version}`;
+
+                    // Input for 'user' and 'channel'
+                    // Initial value is defined as 'undefined' to put the logic at the end.
+                    // Even there is no user and channel it should be at least an empty string, not 'undefined'
+                    let user = undefined;
+                    let channel = undefined;
+
+                    const userInput = await vscode.window.showInputBox({
+                        title: `Editable - Select 'user' for package '${packageInformation}'`,
+                        validateInput: text => {
+                            if (text.includes(" ")) {
+                                return 'Whitespace is not allowed.';
+                            }
+                            else {
+                                return null;
+                            }
+                        }
+                    });
+
+                    user = userInput;
+
+                    // Checking if user cancel the process
+                    // Undefined means that user cancels the process
+                    if (user != "" && user != undefined) {
+                        const channelInput = await vscode.window.showInputBox({
+                            title: `Editable - Select 'channel' for package '${packageInformation}'`,
+                            validateInput: text => {
+                                if (text.includes(" ")) {
+                                    return 'Whitespace is not allowed.';
+                                }
+                                else if (text === "") {
+                                    return "Enter a name for the channel";
+                                }
+                                else {
+                                    return null;
+                                }
+                            }
+                        });
+
+                        channel = channelInput!;
+                    }
+                    // Put channel to empty string if the user is not defined. Both of them have to be existing.
+                    else if (user == "" && user != undefined) {
+                        channel = ""
+                    }
+
+                    // Last step is to put the selected package to editable mode with all the given information
+                    // Make sure that user and channel are string, not 'undefined', and then we can execute the conan API command to add the editable package
+                    if (user != undefined && channel != undefined) {
+
+                        const layout = await vscode.window.showInputBox({
+                            title: `Editable - Select 'layout' for package '${packageInformation}'`
+                        });
+
+                        if (layout != undefined) {
+                            this.conanApi.addEditablePackage(wsChoice.label, packageInformation, user, channel, layout);
+                            vscode.window.showInformationMessage(`Editable package '${packageInformation}' with user '${user}' and channel '${channel}' has been added.`);
+                        }
+
+                    }
+                }
+            }
+            catch (err) {
+                vscode.window.showErrorMessage((err as Error).message);
+            }
+        }
+    }
+
+    private async removeEditablePackage() {
+        try {
+            let editablePackageRecipes = this.conanApi.getEditablePackageRecipes();
+
+            const quickPick = vscode.window.createQuickPick<vscode.QuickPickItem>();
+            let quickPickItems = [];
+            for (let i = 0; i < editablePackageRecipes.length; i++) {
+                quickPickItems.push({
+                    label: editablePackageRecipes![i].name,
+                    description: "",
+                    detail: editablePackageRecipes![i].path,
+                    index: i
+                });
+            }
+            quickPickItems.map(label => ({ label }));
+            quickPick.items = quickPickItems;
+
+            const choice = await vscode.window.showQuickPick(quickPickItems);
+
+            if (choice) {
+                this.conanApi.removeEditablePackageRecipe(choice.label);
+                vscode.window.showInformationMessage(`Editable package ${choice?.label} has been removed.`);
+            }
+        }
+        catch(err) {
+            vscode.window.showErrorMessage((err as Error).message);
+        }
     }
 }